requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dependency-injector==4.32.3
Django==3.2.3
djangorestframework==3.12.4
future==0.18.2
idna==2.10
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
serializers==0.2.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.5
